requirements.txt
Flask
pandas
numpy
joblib
scikit-learn
lightgbm
catboost
xgboost
snowflake-connector-python
python-dotenv
Werkzeug
gunicorn
dotenv
boto3